fix editing on comments
